Okay, index matching ID wont working after thinking it through. Instead I am now using a std map but will replace it with something custom and faster later

+++ SandboxServer/src/ServerSystem.h
@@ -34,7 +34,7 @@ class ServerSystem : public EngineSystem
         bool RemoveClient(u16 ClientID);
         bool RemoveClient(u16 ClientID, std::string Address, u16 Port);
 
-        i32 NumClients() { return m_Clients.size(); }
+        u32 NumClients() { return (u32)m_Clients.size(); }
         //
 
     private:
